Refactor into package per size
