+++ benchmark/src/main/java/com/novoda/sexp/medium/LinkAttributeMarshaller.java
@@ -1,4 +1,4 @@
-package com.novoda.sexp;
+package com.novoda.sexp.medium;
 
 import com.novoda.sexp.marshaller.AttributeMarshaller;
