+++ benchmark/src/main/java/com/novoda/sexp/medium/SexpMediumXmlBenchmark.java
@@ -1,6 +1,9 @@
-package com.novoda.sexp;
+package com.novoda.sexp.medium;
 
 import com.novoda.sax.RootElement;
+import com.novoda.sexp.Instigator;
+import com.novoda.sexp.RootTag;
+import com.novoda.sexp.SimpleEasyXmlParser;
 import com.novoda.sexp.finder.ElementFinderFactory;
 
 import java.util.List;
