+++ benchmark/src/main/java/com/novoda/simple/small/SimpleFrameworkSmallXmlBenchmark.java
@@ -1,4 +1,4 @@
-package com.novoda.simple;
+package com.novoda.simple.small;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
